Git ignore all of the test Berksfile.lock files.
We can't symlink these to the top-level Berksfile.lock because the
relative paths inside are different, and there isn't any easy way for us
to pass a berksfile path to test kitchen. (Ideally we would be able to
pass the -c option to berks, but there doesn't appear to be any way to
do so.)

So these Berksfile.lock files end up having a lot of churn and not being
super useful since they're regenerated every time tests run anyway.