Stop spamming root with cron copies of client.rb.
Running cat in these jobs is important because it ensures that the
command will error if chef-client.rb doesn't exist (which chef-client on
its own doesn't do).

But every time a cron script has output, it gets mailed to root. And we
don't need to accumulate another copy of chef-client.rb in root's inbox
every 10 minutes.