Fix getpglogs shebang line.
We have to use the shim directly for shebangs, using env doesn't work.

The env rbenv exec shebang line doesn't work because on Linux everything
after the first command is treated as a single argument to the command.
This causes env to try to find a script called `rbenv exec ruby`, which
of course doesn't exist.

    #!/usr/bin/env rbenv exec ruby

On an unrelated note, this whole pglogs recipe is unused. Not sure what
the history is here.