[LG-202] Validate CRL signatures
**Why**:
CRLs aren't sent over HTTPs because that requires
validating a certificate. It's turtles all the way
down. So instead, CRLs are signed by the issuing
certificate.

**How**:
We make sure that the CRL is signed by a certificate
we have in-hand. More importantly, it's signed by the
certificate with which the CRL URL is associated.