Move auto EIP config from secrets to git data bag.
**Why**:

Our static IP configuration is private, but not secret. If our public IP
address ranges were disclosed publicly, it would make it easier for an
adversary to understand our network and attack our infrastructure. But
most of this data is already published in whois by ARIN anyway.

https://whois.arin.net/rest/net/NET-34-216-215-0-1.html
https://whois.arin.net/rest/net/NET-34-216-215-128-1.html

Storing private-but-not-secret data in our Citadel S3 buckets alongside
secrets creates hurdles for operations and visibility. The data is not
as visible to team members, and it is harder to determine who changed
what when without the convenience of a git log.

For example, we had an issue in production where I thought I had changed
the production EIP jumphost ranges to a certain range, but actually had
not. This took a long time to discover, and it would have been easier to
troubleshoot if the full team had easy visibility into all of our IP
address ranges in one place.

Our primary argument against storing private-but-not-secret data in
identity-devops has been that we intend to make identity-devops open
source. But this goal has been on the backburner for at least a year,
and we have not made any significant moves toward making it public.

Instead of making identity-devops open source, I propose that we start
moving everything that's currently in identity-devops out into repos
that are already public. This allows us to gradually move cookbooks and
modules into open source repos rather than having to take the giant leap
of flipping identity-devops to public all at once, which would be
difficult since we would need to remove any private data and squash the
entire repo history.

For chef cookbooks, we can move them into our public cookbook repo:
https://github.com/18F/identity-cookbooks/.

For terraform modules, we can move them into our public modules repo:
https://github.com/18F/identity-terraform/.

Also relevant for backstory:
https://github.com/18F/identity-devops-private/issues/1

If we like this approach we may at some point want to fold
identity-devops-private back into identity-devops, since the original
raison d'etre of identity-devops-private was to provide a repo that
would never be made public.

**How**:

- Move most of the static_eip functionality into a standalone static_eip
  cookbook in the public repo.
- Move the static EIP configuration into a data bag,
  private/auto_eip_config.json. This single file contains the EIP ranges
  for all environments.
- The auto_eip_config.json files in S3 will no longer be used.