Add ls-servers --long option and move into module.
Start moving ls-servers functionality into a module, since the CLI code
has gotten way too complex.

Add some more flexible logic around the headers to allow for future
functionality to customize the fields that are selected in the server
listing.

Remove IP addresses from the default headers, and add a human-friendly
uptime, which is easier for humans to parse than the launch time.

Also add availability zone to the ls-servers list, since it's useful to
start building passive awareness of what our story is for high
availability for each server type. This makes it more obvious that we
are only using 2 AZs for most server types.

Add a -l / --long option to ls-servers, which prints all the known
headers.