Add hidden env var to allow scale in during grace.
Sometimes it is useful to be able to scale in servers while some new
instances are still in the health check grace period. This isn't
recommended behavior, so don't add documentation for it, but add
`UNSAFE_ALLOW_INSTANCE_IN_GRACE_PERIOD` as an escape hatch in case
someone wants the relative safety of the scale-in-old-servers script,
but really doesn't want to wait for new instances to be out of the
health check grace period.