Disable dummy jobs by default, no longer needed.
Since we have moved all environments to use the inline activejob queue
adapter, we no longer need to inject dummy jobs into the queues to
perform healthchecks of worker servers, since there are no worker
servers.

Keep the functionality toggled off under an attribute in case we decide
to switch to some kind of queueing system in the future.