Add lifecycle hooks to the IDP ASG provisioning.
This represents a huge improvement for the safety of ASG rollouts.
Auto Scaling Groups have extremely problematic behavior when it comes to
the ASG health check grace period. Instances that are in the grace
period are considered `InService` and therefore `Healthy`, even if the
underlying ELB health check is actually failing. This means that there
are many circumstances where an ASG can terminate the last healthy
instances when there are no other fully bootstrapped healthy instances.

Lifecycle hooks provide a way for ASG instances to be considered
`Pending` until there is an explicit signal to continue to the
`InService` state. The provision.sh script that runs chef bootstrapping
is a natural place to add these hooks.

Add two lifecycle hooks, one for identity-devops-private provisioning
(15 minute timeout), and one for identity-devops provisioning (30 minute
timeout).

- Add lifecycle hook support to provision.sh:
    - Add --asg-name, --lifecycle-hook-name options for specifying a
      lifecycle hook to notify upon provision.sh success/failure.
      When provision.sh completes successfully, it will send the
      CONTINUE result to the hook. If it fails, it instead sends
      ABANDON, which will cause the instance to be terminated.
    - Add --lifecycle-hook-abandon-delay, which causes provision.sh to
      sleep a little bit so a human has time to catch an instance that
      has failed provisioning and is about to be terminated.

- Gzip user data. It was getting too large to fit in the size limit. We
  should probably move provision.sh into the AMI instead.

- Give all servers permissions to complete lifecycle hooks for their
  environment by appending a permission to our default secrets policy.

- Add lifecycle hook options to the flags passed to provision.sh by
  cloud-init.

- Use the new asg_lifecycle_notifications module from identity-terraform
  to add lifecycle hooks to the IDP ASG. Note that due to a terraform
  bug, these hooks might not be effective on the first run when the ASG
  is created for the first time, so we might be introducing some minor
  issues for creating an environment for the first time.