Add lifecycle hooks to ALL autoscaling groups.
Also drop the health check grace period to 0 for all of the autoscaling
groups. Having a health check grace period is dangerous because the ASGs
have braindead behavior when it comes to instances in the grace period.

Instances in the grace period are considered `InService` and `Healthy`,
even if the underlying health checks are failing. This means that an ASG
may terminate the last actually healthy instances if there are new
instances that are still in the grace period but not working.

Thanks to the lifecycle hooks, we no longer need a health check grace
period, because the grace period only starts once an instance enters
state `InService` (i.e. after launch lifecycle hooks have completed).

By this point, provisioning should be complete and instances should be
passing health checks unless something has gone wrong.