Add support for passing TF variables via .tfvars.
For a very long time we have wanted to move away from using environment
variables to pass in terraform variables. Using .tfvars files allows us
to be more explicit and less verbose in specifying our variable
configuration for each environment.

Because .tfvars files are static, we must refactor a few of the places
where we had previously relied upon interpolation in the .sh files. Now
that we are using Terraform 0.10, which supports locals, we can move
variable interpolation into the .tf code itself. This makes the sources
of the defaults somewhat easier to follow.

This also paves the way for moving to a more standard workflow for
running terraform where have a directory per environment and no longer
need to delete .terraform on every run. That is a very desirable end
state because it will allow us to toss out our brittle terraform remote
state configuration in favor of the built-in `terraform init`
functionality present since TF 0.9, it will make terraform runs
significantly faster and less likely to lose data, and it will make it
easier to use standard terraform commands without any wrapper scripts.

If the config for an environment in identity-devops-private sets the
`MOVED_TO_TFVARS` variable, then assume that variables can now be found
in .tfvars files under identity-devops-private:/vars/.

A new script, `get-tfvars-for-env`, looks up the appropriate files in
the vars directory and returns them in the order that they should be
referenced using the `-var-file` option. Terraform uses a
last-write-wins approach to resolving conflicts, which matches our
existing config cascade from base => account global => env specific as
long as we pass them in that order.

Also refactor the way we pass a few different variables as part of
removing interpolation from the variables files:

Also simplify a few parts of configuration, including:
- Set AMIs through a single variable `default_ami_id` or a map from role
  to AMI IDs, `ami_id_map`, which can be used to override the AMIs for
  specific roles. This is much less complicated and verbose compared to
  the previous approach.
- Remove the role-specific `bootstrap_{main,private}_git_ref_<role>`
  variables. In practice these were confusing and hard to use. Set the
  ref for all instances using the standard
  `bootstrap_{main,private}_git_ref`. If we want this functionality back
  we can reintroduce it using a map.