Add a guard in diff-deploy against unexpected cwd.
Because I have diff-deploy on my PATH, I keep running diff-deploy from
my primary identity-devops checkout, even when my CWD is a secondary
identity-devops checkout. This is dangerous because I expect diff-deploy
to be using the git branch and terraform-app config from my CWD, when it
actually uses the config from the checkout on my PATH. Add a sanity
check to guard against this situation and prompt the user to confirm.