Reintroduce per-role branches for main git ref.
The elk and elasticsearch servers had some special variables to
configure the branches they use for provision.sh for identity-devops and
identity-devops-private. This was confusing because the defaults didn't
apply to them, unlike every other server.

Unify this behavior in the same way that we also just did for AMIs.
Create a bootstrap_main_git_ref_default variable that applies to all
servers by default, and create a bootstrap_main_git_ref_map variable
that is a mapping from server role to the branch that should be used for
identity-devops provisioning.

I didn't bother doing the same for identity-devops-private because it's
very rare to want a branch other than master in -private, and even rarer
to need to do so on a per-role basis.