Use a map variable to enable proxy per-role.
Putting the proxy configuration in identity-devops code makes it harder
for us to toggle the proxy on or off because we will need to check out a
different identity-devops branch in order to make changes.

Instead, use a map variable `proxy_enabled_roles` that we pass through
to the bootstrap module. For a role named "rolename", if the map
contains a 1 for "rolename", then the proxy will be enabled. If no entry
is found in the map for that name, try "unknown" as a default. If the
map contains a 0 for the role, then the proxy will not be enabled in the
launch configuration for that role.