Set wait_for_capacity_timeout = 0 on all ASGs.
By default terraform will wait 10 minutes for an auto scaling group to
reach its desired capacity with all instances InService and Healthy.
This is very much not what we want most of the time. Terraform has
extremely bad behavior when interrupted (i.e. potential data loss and
corruption), so it is important that we separate the launch of instances
from the terraform run.

This variable gained more importance recently with the addition of
lifecycle hooks, because if lifecycle hooks are present, then terraform
will wait until instances are fully InService (waiting for the entire
provisioning process to complete), whereas previously this would only
wait for EC2 to have finished creating the unprovisioned instances.