Prepare for TF 0.11 upgrade by fixing outputs.
https://www.terraform.io/upgrade-guides/0-11.html#error-checking-for-output-values

In terraform 0.11, output variables that reference attributes of
conditional resources (i.e. resource that have a count set) must use a
special horrible splat syntax to work around the case where the resource
doesn't exist. Previously terraform would silently return "" if the
resource didn't exist, but in TF 0.11 it enforces that you handle this
"correctly" and checks for errors in output variables.

https://github.com/18F/identity-terraform/commit/62385b497f5b8dba2478be5759d53c1fb2353185