diff-deploy: use interactive apply by default.
Terraform since 0.10 has supported an -auto-approve=false option that
allows you to get an interactive diff. This largely obviates the need
for diff-deploy entirely. Preserve the old functionality of the
landscape gem behind an option.

Switching to the build-in behavior will be much faster since it won't
have to do the annoying `rm -rf .terraform` and re-download of the
modules that we do in `./deploy` by running it twice.