[LG-357] Log untrusted certs to S3
**Why**:
We need to know which certs are failing so we can
better provide user support and pro-actively add
new certificate authorities when appropriate.

**How**:
We put the certificate text and PEM content in an
object in S3. We can configure the S3 bucket so
that the server only as write access and not read
access.