Port asg-recycle.sh to ruby in cloudlib.
This adds a lot of lines of code, which isn't the greatest, but it's
much easier to read and maintain in ruby. It was also much easier to add
some nicer functionality like checking to see if we are already in the
middle of a recycle and prompting whether to continue if so.