Run 3 proxy servers by default to match AZs.
Since we run outbound proxy servers in 3 AZs, the auto recycle
functionality doesn't work correctly unless the number of instances
matches the number of AZs.

When an ASG scales in servers, it applies any termination policy such as
OldestInstance first on a per-zone basis to any zones with more
instances than the rest. If all zones have the same number of instances
it then applies the policy globally.