Compile passenger module for all ruby versions.
Passenger has a native module that it expects to exist to improve
operation. This is specific to each ruby version, so if the global
version of ruby that passenger uses is different from the version of
ruby used by an individual app, passenger will attempt at runtime to
compile the module.

This fails because we deliberately don't give passenger permissions to
write files.

Instead, when we are setting up passenger for the first time, also
compile the native module for all available rbenv ruby versions.

The directory where passenger compiles and looks for the binaries is
configured with PASSENGER_NATIVE_SUPPORT_OUTPUT_DIR.