Switch from SPF hardfail to softfail for SES.
Previously we had a hardfail SPF policy (-all). Switch to using a
softfail policy instead (~all). This has little or no impact for most
email providers, which will either outright ignore SPF failures or else
treat them as a small input signal going into spam detection algorithms.

This is because SPF does not handle forwarded email, so email receivers
tend to use other mechanisms such as reputation and DKIM signatures to
detect inauthentic messages.

We use a strict DMARC reject policy, which for compliant providers means
that a message *must* pass one of DKIM or SPF checks, or it will be
rejected. This provides a much stronger email authenticity guarantee
than SPF.

For the vast majority of email providers, this change will make no
difference. However, there are a handful of email providers that reject
SPF failing messages when there is a hardfail policy even for messages
that pass DKIM signatures. This prevents users from forwarding email
through other providers. Moving instead to a softfail policy will
improve email deliverability for the very small set of users in this
situation.