Avoid automatic switching to a multipart upload with boto __Why__
* The `x-amz-server-side-encryption` header is not added to the `Initiate
Multipart Upload request` when boto automatically switches to multipart file
uploads (due to large file size). This appears to be fixed in s3, which at a
time did not support kms encryption with multipart uploads:
https://github.com/boto/botocore/issues/471#issuecomment-147514394
That appears to be fixed now:
https://docs.aws.amazon.com/AmazonS3/latest/dev/KMSUsingRESTAPI.html

Still, it seems our logstash-output-s3 plugin is using an older version of boto
or is not correctly telling boto to pass the correct SSE headers to boto. As a
work around we can configure boto to avoid multipart uploads by increasing the
size threshold.

It may be worth looking into the logstash-output-s3 code to see if this can be
fixed since it may be as easy as updating to a newer version of boto.