Task to fetch new issuing certs
**Why**:
When a user tries to use a PIV/CAC with their account and we
don't have the issuing cert, we fail to verify the PIV/CAC
cert and log the cert for further investigation. Once we
determine that we should support the PIV/CAC, we find the
issuing certificate and bundle it for the next deploy.

**How**:
The script fetches the currently deployed set of issuing certs,
the current set of logged certs, and sees if any issuing certs
are missing from the deployed bundle. If so, it tracks them down
and prints them to STDOUT for inclusion as long as they chain up
to a trusted root.