Log OCSP responses (#55)
**Why**: Some users are reporting that their PIV/CAC is coming back as
revoked even though they/we have reason to believe they aren't. We need
to log the OCSP responses so we can see what's causing the system to
mark the PIV/CAC cert as revoked.

**How**: If the OCSP response indicates that it is not successful OR
that the PIV/CAC cert is revoked, we drop the response along with some
textual content into the same S3 bucket into which we log unverifiable
certs.