Add an unused duplicate copy of the obproxy NLB.
Create a copy of the outboundproxy NLB in the public subnets. We will
run this in parallel in order to make a graceful transition without
downtime. Once this has been rolled out to all environments, we can cut
over by repointing the DNS name `obproxy.login.gov.internal` at the new
NLB in the public subnets.

We are making this move in response to a bizarre NLB bug (design flaw?)
where if the source and dest sides of a connection are in the same
subnet, but the NLB is in a different subnet, then the connection fails.