Chdir to identity-devops for bundler/setup.
Bundler gets confused when you run any of the bin scripts from outside
of the identity-devops directory. This is because it uses the current
working directory to find the Gemfile.

Temporarily change directories to identity-devops when requiring
`bundler/setup` in order to work around this.

This allows all of the cloudlib scripts to be added to the PATH and run
from anywhere.