Add OIDC cert and key
**Why**: Historically, the OIDC and SAML implementations in the IdP have
both used `certs/saml.crt` and `keys/saml.key.enc` where they needed a
private key, public key, or certificate. This has led to some confusion
when making changes to those files.

This commit is part of an effort to simplify things by copying the SAML
files into new files that the OIDC implementation can use. This allows
us to operate on the SAML or OIDC IdP code without having to worry
about the side-effects in the other part of the code.

One place where this doesn't make sense and their is room for a refactor
in the future is that OIDC does not have a need for a certificate, just
a public key, so keeping a certificate there doesn't make the most
sense.