Run aideinit as final cloud-init step.
This makes aideinit still appear in cloud-init-output.log, which is
useful, but it avoids the problem of running aideinit twice. Aideinit
takes a long time to run when there are a lot of files on the system.

Provision.sh is run twice during bootstrapping, once for
identity-devops-private, and once for identity-devops. (This
architecture is a little weird in retrospect, but it seemed like a good
idea at the time.)