provision.sh: Set proxy vars in /etc/environment.
This is necessary for test kitchen to work properly. In the normal chef
run, provision.sh sets local environment variables when it runs chef,
but test kitchen doesn't know that it needs to do this when running chef
for tests.

To address this, set the appropriate environment variables in
/etc/environment. This ought to cause new sessions to incorporate the
appropriate proxy variable automatically, even before the first chef run
occurs.