Set apt proxy in cloud-init base template.
This is not normally needed on instances because apt is invoked from a
context where pam_env reads from /etc/environment to set the http_proxy
environment variables. But this doesn't necessarily happen in the first
run under kitchen-ec2 for testing. So it's a useful fallback to ensure
that apt is using the expected proxy configuration.