Move proxy config into chef_proxy_env cookbook.
It turns out that a number of the test failures in CI were because Test
Kitchen does not use the chef-client.rb in this repo. It instead
generates its own chef-client.rb, and there doesn't appear to be any way
to set arbitrary variables inside this config.

As a result, the only foolproof way we have to set the chef HTTP proxy
variables is from inside a chef run using a cookbook.

Create a new cookbook, `chef_proxy_env`, which uses its default recipe
to look for proxy configuration from files at a specified location.
These files must exist or the recipe will raise an exception, but if
they are empty then no proxy will be used. For Login.gov, we look for
these proxy config files in `/etc/login.gov/info/`.

This resolves some long-running problems with chef runs in CI not using
the expected proxy settings. We also ought to be able to unwind some of
the workarounds that explicitly set proxies in a few places.