Switch VPC endpoints to use base SG for access.
Previously we would list each security group for each server role in
every single VPC endpoint. This was extremely duplicative. We could add
a variable for this, but better yet we can switch to the relatively new
"base" security group.

All instances are expected to be added to the base security group, which
grants certain base network access necessary for all instances.

Add the base security group to jumphosts, which did not have it
previously.

Because jumphosts won't be in the base group until their next recycle,
keep jumphosts in the ingress rules for the VPC endpoints until this
release is fully rolled out (then the jumphost group can be removed).

For KMS, continue to allow only idp (plus the new migration servers),
since most servers do not need to hit KMS directly.

Change all other VPC private endpoints to use the base security group as
the ingress allow list.

Also add the base security group to the security groups assigned to
instances created by test kitchen. We have to duplicate the config in
kitchen.yml that is specified in Terraform for the ASG when we create
these instances outside of the launch template. (Although maybe there's
something we could do to look this up more dynamically.)