Update instance_certificate cookbook for Chef 14.
Fix some deprecation warnings around the lack of
current_resource/new_resource references when accessing properties.

Also make the generate resource actually able to load existing
certificates in a chef-native way, and use a pass-through to the
underlying file resources to let chef decide when it needs to write out
changes to the cert/key files.

This shortens the generated certificate subject names, which will cause
certificates to be regenerated on instances that pull in the new
cookbook.