Refactor tf-deploy to accept new subdir layout.
**Problem statement**:

Since the inception of Login.gov's infrastructure, we have used a shared
terraform directory layout where a single directory of terraform config
(such as "terraform-app") is used to deploy to multiple different
environments and accounts.

In the very early days, the env-specific variables were supplied from
untracked local files. We migrated away from that practice to tracking
all remaining env-specific config variables in identity-devops-private.
This separation was intended to allow us to open source identity-devops
without exposing any private-but-not-secret data.

However, we ultimately have abandoned plans to open source
identity-devops due to the difficulty of splitting out all the private
data.

Moreover, the shared directory strategy has always been seriously
problematic because it requires blowing away and reconsituting the
`.terraform` directory before each terraform run. This makes it more
difficult to run terraform commands standalone without our `deploy`
scripts, and also increases the risk of catastrophic statefile damage.
Recent versions of terraform use the local .terraform directory to
determine when the remote state file needs a version upgrade. So by
blowing away the directory, we lost that information.

We made significant incremental improvements to that strategy by
treating .terraform as a symlink to persistent directories under
`./deploy/`. This allowed us to reap the benefits of the persistent
local state version tracking in newer terraform versions, and reduced
the risk of loss. However, this approach was still not very good because
it requires a rube goldberg machine of scripts that source environment
variables in shell scripts in identity-devops-private.

We made significant imcremental improvements to _that_ strategy by
moving most environment variables out of shell scripts and into
`.tfvars` files in identity-devops-private, which makes the
configuration more declarative and less of a mess. But much of the rube
goldberg machine still remains.

**New .terraform layout**:

The new strategy uses separate subdirectories for each environment. This
mirrors the approach that I've taken when using Terraform on other
projects. This is a much simpler and more reliable approach because most
of the dependencies are explicitly encoded in the directory layout and
in terraform config files, not a maze of shell scripts.

In the new layout of terraform config, most configuration will go into a
module shared between environments. This has the advantage of explicitly
encoding all of the variables that an environment must supply. It means
that any unexpected or undefined variables will be treated as errors
rather than ignored, which reduces potential for mistakes. This also
will make it easier to open source individual modules in the future,
because the configuration will already be modularized, meaning no
statefile surgery would be needed.

So common configuration might go in `terraform-sms/module/*.tf`.

When creating a new environment, we will create a directory like
`terraform-sms/sandbox/`. This directory primarily contains `main.tf`,
which instantiates the module from `../module/` and provides any
environment-specific variables, as needed.

Other environments that use the same module code can go next to this one
and supply their own environment-specific variables. All variables will
be self-contained inside this directory tree. No terraform variables are
sourced from shell scripts, and nothing is loaded from
identity-devops-private.

To preserve compatibility with the `tf-deploy` family of scripts and to
ease creation of the remote state S3 bucket and dynamodb table, we will
still use a shell script with shell variables, `<env>/env-vars.sh`.

For example, `terraform-sms/sandbox/env-vars.sh` contains the
information needed to bootstrap the state bucket and terraform state.
None of these variables are provided to terraform runs directly, they
only configure our bootstrapping scripts. This means that after the
initial bootstrap, you ought to be able to cd to this subdirectory and
run `terraform` by itself with no wrapper scripts.

**How**:

We have a bit of a UX parsing problem with the `tf-deploy` script usage,
since it expects to receive a TF_DIR and ENV separately.

To differentiate between old shared identity-devops-private style
directories and the new separate module style directories, we will adopt
a convention of supplying the full directory path
`<terraform-foo>/<env>` when we want the new style.

So old style usage will continue to look like this:

    tf-deploy dev terraform-app plan

While new module / subdirectory style usage will look like this:

    tf-deploy terraform-sms/sandbox plan

I'm not 100% sure this is the least confusign way to handle this, so
suggestions here are welcome.

Implementation steps:

- Create a symlink from `deploy` to `bin/tf-deploy` so that users can
  seamlessly continue writing `./deploy`.
- Implement processing of the new subdirectory style layout in
  `tf-deploy`. If a TF_DIR contains a `/`, we assume it's the new
  subdirectory style, otherwise we use the old style.
- Add new `env-vars.sh` for terraform-sms/sandbox/.
- Refactor `configure_state_bucket.sh` to handle both --module-style and
  --shared-style .terraform layouts.