Add a script to import TF remote state resources.
We have a bootstrapping problem when attempting to manage TF remote
state with terraform. Terraform needs the remote state bucket to exist
before the first run, but we also want to manage this bucket with
terraform.

As a result, we have configure_state_bucket.sh create the remote state
bucket automatically. But then we want to import this resource into
terraform for management.

Add a `--import-remote-state` option to `tf-deploy`, which will handle
performing the import for you. Also reference the new `state_bucket`
module from the `terraform-master` directory, which makes this
management much easier.

When you run `tf-deploy --import-remote-state <tf_dir>`, it will run
terraform import on the S3 remote state bucket and DynamoDB remote state
lock table. It expects to find a variable like this in `env-vars.sh` in
order to know the terraform address under which to import the resources:

    # used by `tf-deploy --import-remote-state` to know TF address for importing
    ID_state_module_prefix=module.main.module.tf-state