Replace the passenger restart hack with prewarm.
Historically, we seemed to need the passenger restart hack to ensure
that we had a running passenger at the end of a chef run. I'm pretty
sure this hasn't actually been necessary in ages.

But recently we have started seeing a phenomenon where servers that
successfully bootstrap are immediately terminated after failing health
checks. I suspect that this is because it is taking too long for the
servers to start up passenger processes when the ELB starts sending them
health checks.

We can jump start this process by sending an HTTP request to localhost
at the very end of the chef run. This should prewarm the HTTP server and
prevent the ELB from concluding that it is unhealthy.