Remove a whole bunch of unused stuff.
Remove a bunch of config and documentation related to worker servers,
which no longer exist.

Also remove the identity-jenkins cookbook, which is unused. We
ultimately moved to autoscaling groups instead of rolling out Jenkins in
production.

Remove some old recipes that were necessary when we ran a chef server
and used terraform instance provisioners to do bootstrapping.