Explicitly set RDS engine version on replica.
Evidently RDS cannot upgrade the primary until all read replicas have
been upgraded. And when we don't explicitly specify an engine version,
Terraform doesn't know that it should upgrade these.

    * aws_db_instance.idp: Error modifying DB Instance: DBUpgradeDependencyFailure: One or more of the DB Instance's read replicas need to be upgraded

Also split out the variables used to control these. It's insufficient to
bump both of the versions at once, because Terraform sees that the
replica depends on the primary, so it thinks that the primary must be
changed first.

In reality, the replica must be upgraded first due to RDS requirements.
So we can only accomplish this in two phases by first upgrading the
replica.