additional values needed for DB + making DB build nonconditional
In testing, I discovered that there were two parameters still needed
in order to build the worker DB -- the database username and password.
For the main DB, those variables are left blank in this repo, and
declared with defaults in `base.tfvars` in `identity-devops-private`.
I made new variables for these, specific to the worker DB, and
declared those over in that repo.

In discussing with @pauldoomgov, it also looks like we're going to
want worker hosts present everywhere, by default. Thus, I updated the
defaults for Desired and Minimum counts to 1, and removed the `count`/
conditional from the DB so that it's always created.

Lastly, I moved the DB resource block into `worker.tf` -- since the
other IdP resources (aside from the DB) are all within `idp.tf`,
this seemed like the more logical place for the DB to go.