Add tf-deploy shell (#4065)
`tf-deploy` takes time to setup on each call.   This adds a rudimentary
`shell` option to allow running a series of `terraform` commands
in succession on the same state.

It also adds `bin/oneoffs/export-state-ids.py` which enumerates the
resources and their ID values for a given state file.   This may
be useful for troubleshooting, or for modeling a set of imports
when some absolute chucklehead deletes someone's sandbox state file.

(Editor's note: @pauldoomgov is said chucklehead.)