new SSM ecosystem: docs, scripts, policies, et al (#4392)
* updated gsa-username ssm doc + functionality in ssm-instance

* configuring to use session tags for passing username

* use identity-terraform/ssm module instead of private-repo module

* adding id-uuid-lookup script + SSM doc + role updates

* ssm vs ec2 permissions/conditions + adding ec2_username

* updates to IAM/common-related scripts

* adding work-restart command + commenting SupportEng for now

* keep old resources intact for overlapping environments

* updating identity-terraform gitsha + creating SupportEngineer role

* updates to ssm-instance + update condition for sts:TagSession

* only allow ssm:StartSession if an SSM document is provided with --document-name

* commenting in some TODO resources RE: enforcement

Optimally, there should be a way to get a list of all of the roles NOT in var.ssm_access_map, and attach
a policy that DENIES ssm:StartSession access to them (since they are not in said variable) -- but ONLY
when var.ssm_access_map actually exists. These commented-out resources are one way to do it, but in the
interest of keeping code DRY -- and the commit this PR is attached to limited to its original scope --
they're simply TODOs for now, to preserve the work and ideas presented herein.

* hack in ssm-instance for gitlab access

* wrong gitsha