Add a script for copying a sandbox cloudwatch dashboard to a terraformed dashboard (#5065)
This commit adds a script which takes a sandboxed dashboard and copies the config over to be a terraformed dashboard. This makes it easier to work on dashboards in AWS and copy them over to terraform.

The script handles things like setting the appropriate variables for the region or for the environment.