Use tuples in the failed resolutions attributes dashboard (#5421)
The failed resolution attributes dashboard was previously showing failures by attribute. This does not tell the entire story. It is useful to know if users are failing with multiple attribute combinations, e.g. DOB and SSN.

This commit changes the dashboard to show the attributes that fail in tuples so that combintations of failed attributes are revealed.